Update rem_carrier_phase and minor changes

+++ src/algorithms/tracking/gnuradio_blocks/dll_pll_veml_tracking.h
@@ -151,7 +151,7 @@ private:
     boost::circular_buffer<double> d_cp_history;
     // remaining code phase and carrier phase between tracking loops
     double d_rem_code_phase_samples;
-    double d_rem_carr_phase_rad;
+    float d_rem_carr_phase_rad;
 
     // PLL and DLL filter library
     Tracking_2nd_DLL_filter d_code_loop_filter;
